Ya esta funcionando bien :D Haga porfa el documento, ponga sysouts para que se vea chevre y se entienda un poco el funcionamiento del programa, (aunque lo deje bastente entendible) y nose... haga lo que falta. Desde este punto me desentinedo de infracomp mande el trabajo ud tambien porfa.
Ahora... Commit and Push bitch!!

+++ infracom.caso1/Cliente.java
@@ -14,7 +14,8 @@ public class Cliente implements Runnable
 		mensajes = new Mensaje[cantidad];
 		for ( int i = 0; i < cantidad; i++)
 		{
-			mensajes[i] = new Mensaje(i);
+//			mensajes[i] = new Mensaje((int) (Math.random()*100));
+			mensajes[i] = new Mensaje((int) i);
 		}
 	}
